refactor(goals): extract managed state engine
