+++ .agents/zzzops/zzzops.py
@@ -39,6 +39,13 @@ _feedback = importlib.util.module_from_spec(_FEEDBACK_MODULE_SPEC)
 sys.modules[_FEEDBACK_MODULE_SPEC.name] = _feedback
 _FEEDBACK_MODULE_SPEC.loader.exec_module(_feedback)
 
+_GOALS_MODULE_PATH = Path(__file__).with_name("goals.py")
+_GOALS_MODULE_SPEC = importlib.util.spec_from_file_location("zzzops_goals", _GOALS_MODULE_PATH)
+assert _GOALS_MODULE_SPEC and _GOALS_MODULE_SPEC.loader
+_goals = importlib.util.module_from_spec(_GOALS_MODULE_SPEC)
+sys.modules[_GOALS_MODULE_SPEC.name] = _goals
+_GOALS_MODULE_SPEC.loader.exec_module(_goals)
+
 PROJECT_SCHEMA_VERSION = _policy.PROJECT_SCHEMA_VERSION
 PLAN_SCHEMA_VERSION = 1
 POLICY_SCHEMA_VERSION = _policy.POLICY_SCHEMA_VERSION
@@ -91,6 +98,7 @@ MACHINERY_PATHS = (
     ".agents/zzzops/policy.py",
     ".agents/zzzops/reservation.py",
     ".agents/zzzops/feedback.py",
+    ".agents/zzzops/goals.py",
     ".agents/zzzops/.gitignore",
     ".agents/zzzops/INSTALL_MANIFEST",
     ".agents/zzzops/templates/project-goals",
@@ -124,6 +132,13 @@ parse_reservation_description = _reservation.parse_reservation_description
 normalize_resources = _policy.normalize_resources
 normalize_resource_policy = _policy.normalize_resource_policy
 exclusive_resources = _policy.exclusive_resources
+
+parse_managed_goal = _goals.parse_managed_goal
+validate_managed_goal = _goals.validate_managed_goal
+render_managed_goal = _goals.render_managed_goal
+goal_needs_human = _goals.goal_needs_human
+validate_github_issue_goal = _goals.validate_github_issue_goal
+github_goal_record = _goals.github_goal_record
 
 execution_reports_enabled = _feedback.execution_reports_enabled
 execution_report_directory = _feedback.execution_report_directory
@@ -464,164 +479,6 @@ validate_project_state = _policy.validate_project_state
 validate_project_artifacts = _policy.validate_project_artifacts
 validate_policy = _policy.validate_policy
 policy_blockers = _policy.policy_blockers
-
-
-def parse_managed_goal(text: str, issue_number: int | None = None) -> dict[str, Any] | None:
-    pattern = re.compile(
-        re.escape(GOAL_BLOCK_START) + r"\s*\n(.*?)\n" + re.escape(GOAL_BLOCK_END),
-        re.DOTALL,
-    )
-    match = pattern.search(text)
-    if not match:
-        return None
-    try:
-        goal = json.loads(match.group(1))
-    except json.JSONDecodeError as exc:
-        raise ValueError(f"Invalid managed goal JSON: {exc}") from exc
-    errors = validate_managed_goal(goal, issue_number)
-    if errors:
-        raise ValueError("Invalid managed goal: " + "; ".join(errors))
-    return goal
-
-
-def validate_managed_goal(goal: Any, issue_number: int | None = None) -> list[str]:
-    if not isinstance(goal, dict):
-        return ["managed goal must be an object"]
-    errors = []
-    unknown = sorted(set(goal) - GOAL_FIELDS)
-    if unknown:
-        errors.append("unknown fields: " + ", ".join(unknown))
-    if goal.get("schema_version") != GOAL_SCHEMA_VERSION:
-        errors.append(f"schema_version must be {GOAL_SCHEMA_VERSION}")
-    required_text = ("status", "priority", "value", "difficulty", "confidence", "next_action")
-    for field in required_text:
-        if not text_present(goal.get(field)):
-            errors.append(f"{field} is required")
-    for field, allowed in (
-        ("status", GOAL_STATUSES), ("priority", GOAL_PRIORITIES), ("value", GOAL_VALUES),
-        ("difficulty", GOAL_DIFFICULTIES), ("confidence", GOAL_CONFIDENCES),
-    ):
-        if text_present(goal.get(field)) and goal[field] not in allowed:
-            errors.append(f"{field} is invalid")
-    for field in ("depends_on", "blockers", "evidence"):
-        if not isinstance(goal.get(field), list):
-            errors.append(f"{field} must be a list")
-    parent = goal.get("parent")
-    if parent is not None and (not isinstance(parent, int) or isinstance(parent, bool) or parent < 1):
-        errors.append("parent must be null or a positive issue number")
-    dependencies = goal.get("depends_on")
-    if isinstance(dependencies, list):
-        if any(not isinstance(item, int) or isinstance(item, bool) or item < 1 for item in dependencies):
-            errors.append("depends_on entries must be positive issue numbers")
-        if len(set(dependencies)) != len(dependencies):
-            errors.append("depends_on entries must be unique")
-        if issue_number is not None and issue_number in dependencies:
-            errors.append("depends_on cannot contain the current issue")
-    if issue_number is not None and parent == issue_number:
-        errors.append("parent cannot be the current issue")
-    try:
-        normalize_resources(goal.get("resources", []))
-    except ValueError as exc:
-        errors.append(str(exc))
-    blockers = goal.get("blockers")
-    if isinstance(blockers, list):
-        for index, blocker in enumerate(blockers):
-            if not isinstance(blocker, dict):
-                errors.append(f"blockers[{index}] must be an object")
-            elif blocker.get("status") == "open" and blocker.get("category") not in BLOCKER_CATEGORIES:
-                errors.append(f"blockers[{index}].category is invalid or missing")
-    if not isinstance(goal.get("revision"), int) or isinstance(goal.get("revision"), bool) or goal.get("revision", 0) < 1:
-        errors.append("revision must be a positive integer")
-    implementation = goal.get("implementation")
-    if implementation is not None:
-        if not isinstance(implementation, dict):
-            errors.append("implementation must be an object")
-        else:
-            for field in ("branch", "base", "target", "pr"):
-                if field not in implementation or (implementation[field] is not None and not text_present(implementation[field])):
-                    errors.append(f"implementation.{field} must be null or non-empty text")
-            review = implementation.get("review")
-            if not isinstance(review, dict) or review.get("status") not in {"not_started", "pending", "approved", "changes_requested"}:
-                errors.append("implementation.review.status is invalid")
-            elif "checkpoint" not in review or (review["checkpoint"] is not None and not text_present(review["checkpoint"])):
-                errors.append("implementation.review.checkpoint must be null or non-empty text")
-    return errors
-
-
-def render_managed_goal(goal: dict[str, Any], body: str = "", issue_number: int | None = None) -> str:
-    errors = validate_managed_goal(goal, issue_number)
-    if errors:
-        raise ValueError("Invalid managed goal: " + "; ".join(errors))
-    block = f"{GOAL_BLOCK_START}\n{json.dumps(goal, ensure_ascii=False, sort_keys=True, separators=(',', ':'))}\n{GOAL_BLOCK_END}"
-    pattern = re.compile(
-        re.escape(GOAL_BLOCK_START) + r"\s*\n.*?\n" + re.escape(GOAL_BLOCK_END),
-        re.DOTALL,
-    )
-    if pattern.search(body):
-        return pattern.sub(lambda _match: block, body, count=1)
-    separator = "\n\n" if body and not body.endswith("\n\n") else ""
-    return f"{body}{separator}{block}\n"
-
-
-def goal_needs_human(goal: dict[str, Any]) -> bool:
-    return any(
-        isinstance(blocker, dict)
-        and blocker.get("status") == "open"
-        and blocker.get("category") in BLOCKER_CATEGORIES
-        for blocker in goal.get("blockers", [])
-    )
-
-
-def validate_github_issue_goal(issue_number: Any, title: Any, body: Any) -> list[str]:
-    errors = []
-    if not isinstance(issue_number, int) or isinstance(issue_number, bool) or issue_number < 1:
-        errors.append("issue_number must be a positive integer")
-    if not text_present(title):
-        errors.append("title is required")
-    elif REDUNDANT_GOAL_TITLE_PREFIX.match(title):
-        errors.append("title must not contain a redundant ZzzOps goal ID")
-    if not isinstance(body, str):
-        return errors + ["body must be text"]
-    human = body.split(GOAL_BLOCK_START, 1)[0].lstrip("\ufeff\r\n ")
-    if not human:
-        errors.append("human-readable content must precede managed state")
-    elif human.startswith("---"):
-        errors.append("human-readable content must not start with rendered frontmatter")
-    elif not human.startswith("## "):
-        errors.append("human-readable content must start with a section heading")
-    try:
-        goal = parse_managed_goal(body, issue_number if isinstance(issue_number, int) else None)
-    except ValueError as exc:
-        errors.append(str(exc))
-    else:
-        if goal is None:
-            errors.append("managed goal block is required")
-    return errors
-
-
-def github_goal_record(issue: dict[str, Any]) -> dict[str, Any]:
-    number = issue.get("number")
-    body = issue.get("body") or ""
-    goal = parse_managed_goal(body, number)
-    if goal is None:
-        raise ValueError("managed goal block is required")
-    errors = validate_github_issue_goal(number, issue.get("title"), body)
-    if errors:
-        raise ValueError("; ".join(errors))
-    digest_source = "\0".join((issue.get("title") or "", body, issue.get("updated_at") or ""))
-    return {
-        "key": number, "title": issue["title"], "status": goal["status"],
-        "priority": goal["priority"], "value": goal["value"], "difficulty": goal["difficulty"],
-        "confidence": goal["confidence"], "parent": goal["parent"],
-        "depends_on": goal["depends_on"], "claim": goal["claim"], "resources": goal.get("resources", []),
-        "needs_human": goal_needs_human(goal),
-        "blocker_categories": sorted({blocker["category"] for blocker in goal["blockers"] if blocker.get("status") == "open"}),
-        "next_action": goal["next_action"], "revision": goal["revision"],
-        "digest": hashlib.sha256(digest_source.encode("utf-8")).hexdigest(),
-        "updated_at": issue.get("updated_at"), "implementation": goal.get("implementation"),
-        "labels": sorted(label["name"] for label in issue.get("labels", []) if isinstance(label, dict) and text_present(label.get("name"))),
-        "state": issue.get("state"), "url": issue.get("html_url"),
-    }
 
 
 def validate_goal_transition(transition: Any, issue_number: int) -> list[str]:
@@ -1575,6 +1432,9 @@ def nonempty(value: Any) -> bool:
 
 def text_present(value: Any) -> bool:
     return isinstance(value, str) and bool(value.strip())
+
+
+_goals.configure_entrypoint(normalize_resources=normalize_resources, text_present=text_present)
 
 
 def nonempty_list(value: Any) -> bool:
